feat: add plan-backed monthly resource limits
Persist four monthly resource amounts on Plan Rows and expose them
with current usage through the aggregate account limits endpoint.

BREAKING CHANGE: replace per-agent limit GET endpoints with
GET /v1/accounts/{account_id}/limits and replace legacy monthly Plan
fields with four dimension-specific amounts.

Refs GOL-528

+++ integration-tests/tests/memory_billing.rs
@@ -97,7 +97,7 @@ mod tests {
             .await
             .get_account_usage(&user.account_id.0, None)
             .await?;
-        Ok(usage.usage.durable_storage_gb_month * BYTE_SECONDS_PER_GB_MONTH)
+        Ok(usage.usage.durable_storage_gb_month * BYTE_SECONDS_PER_GB_MONTH as f64)
     }
 
     #[test]
